+++ source/system-buttons/include/Powerdown.php
@@ -1,5 +1,5 @@
 <?PHP
-/* Copyright 2012-2020, Bergware International.
+/* Copyright 2012-2023, Bergware International.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License version 2,
@@ -10,10 +10,15 @@
  */
 ?>
 <?
-switch ($_POST['cmd']) {
-  case 'shutdown': $cmd = 'poweroff'; break;
-  case 'reboot'  : $cmd = 'reboot'; break;
+if (isset($_POST['cmd'])) {
+  switch ($_POST['cmd']) {
+    case 'shutdown': $cmd = 'poweroff'; break;
+    case 'reboot'  : $cmd = 'reboot'; break;
+    default        : $cmd = ''; break;
+  }
+  if ($cmd) {
+    exec("/sbin/$cmd 1>/dev/null 2>&1 &");
+    echo $_POST['cmd'];
+  }
 }
-exec("/sbin/$cmd 1>/dev/null 2>&1 &");
-echo $_POST['cmd'];
 ?>
